Move to 'Azure.Identity' v1.14.2 and refactor the telemetry library (#404)
Getting access token from Azure PowerShell fails with the error `"PowerShell did not return a valid response."`. This turns out to be an issue with the `Azure.Identity` library. After moving to the latest v1.14.2, it works again.

However, moving `Azure.Identity` to v1.14.2 causes a dependency conflict with the `Microsoft.ApplicationInsights.WorkerService` package, so this change also refactors the telemetry code a bit to depend on the  `Microsoft.ApplicationInsights` package instead, which is lightweight and much more appropriate for this application.